The game engine, editors and game are now (almost) completely separate from each other. --
The editor runs in its own ScriptingEngine (Lua state), which is separate from the game's ScriptingEngines (Lua states).
The editor's ScriptingEngine runs in the main thread.

The game has a 'logic' ScriptingEngine and a 'render' ScriptingEngine.
The 'render' ScriptingEngine runs in the main thread and the 'logic' ScriptingEngine runs in its own thread.


Former-commit-id: 777ce578e7ccabe45c85c253dd43e35ce8a045bb [formerly be0bc7e9446cdba981c7b9b39f096c5bb4dfc218]
Former-commit-id: 4bb2f2f5f07d8226303195d2e8f89c1240de5464