Basic operators (-. +. * and /) can now be accessed in Lua via GLM.__add, GLM.__sub, GLM.__mul and GLM.__div.
To make them work as intended, I still need to write a bit of Lua code like this:
getmetatable(GLM.vec2()).__mul = GLM.__mul

The above makes the * operator work for vec2, but the same has to be done for all operators and all classes.

Another way to make operators work, is to extend the classes with new operators, but this only works one way (e.g. vec2 * scalar works, but scalar * vec2 doesn't work).
Because of this, the above operator functions and a big of Lua code is the only way to make operators work properly.


Former-commit-id: 6cc0152a49decb2038a57afdd2fd0ed1aa6bd057 [formerly 179a3877a96414bd4f8e491bf423969019870a78]
Former-commit-id: 8c23ae84c6786d22fc50380973e540ef1c136968