This is mostly an 'archive' commit.
I added a second template parameter to the Container class, which specifies the std container to use in the Container.
It works, but the std::set and std::list implementations of Container are so different, that the second template parameter is pretty much useless.
Not to mention that std::set and std::list will probably be the only Container implementations ever used and using a template for just 2 different versions is also kind of pointless.

So, I'm going back to Container and OrderedContainer, which will only share a base class.


Former-commit-id: 66b4d0521ed2adbc331083e87c17d11bdeb2466b [formerly a7ce2f49df6d24bb68a7629d1b759a88a02d757d]
Former-commit-id: b99142924d0f08d2e3ea5387537eecbb066c23de